feat(32-5 T2): extract agentic tool loop into InlineToolLoopRunner (verbatim, no fork)
Moved AgenticToolLoop + CallAnthropic/OpenAiMultiTurn + the multi-turn body
builders + Anthropic/OpenAI response parsers + the single-turn CallAnthropic
Messages/CallOpenAiCompatible (used by the compaction summarizer) + LoadProvider
Config/LoadProviderConfigWithKeyAsync + GetDefaultModel + Truncate VERBATIM into
new InlineToolLoopRunner : IInlineToolLoopRunner under Tamma.Activities/LlmCall.
The loop body is byte-for-byte identical after only two sanctioned mechanical
substitutions (context.WorkflowExecutionContext.Id -> workflowInstanceId,
context.CancellationToken -> cancellationToken); 424 loop lines verified identical.

CallLlmInlineActivity (1592 -> 520 LOC) now delegates LOCALLY to the runner: the
tool-loop branch calls ToolLoop.RunAsync; SingleTurnCall + GetDefaultModel calls
route through the runner's moved helpers. Constructor signature preserved (added
an optional trailing InlineToolLoopRunner? param) so the [JsonConstructor] and the
4-arg test ctors still compile. The runner is DI-registered in ElsaServer
Program.cs (IInlineToolLoopRunner + concrete, transient) and runs in-engine — DI
relocation to Tamma.Api is deferred to T4/T6.

AC4 no-fork: exactly one AgenticToolLoop definition + one InlineToolLoopRunner
class (both in the runner); the activity contains no loop body. CallLlmInline
ActivitySanitizationTests is byte-for-byte unchanged (regression net, 12 green).
The loop-specific tests (AgenticToolLoopTests/IntegrationTests, CallLlmInline
CredentialTests, EngineProviderCredentialWiringTests) are lifted to exercise the
runner (no assertions weakened). New InlineToolLoopRunnerTests drives RunAsync
end-to-end (multi-turn, tool execution, token totals, turns, exhausted, tool-
output sanitization, preserved http status). LlmCall suite: 350 green. Engine +
API build clean (0 errors). No migration/snapshot change.

InlineToolLoopResult keeps the existing 4-tuple shape: InputTokens/OutputTokens
are the loop's cumulative prompt/completion totals (already folded onto the
response); Turns/Exhausted as before; ToolCalls is the documented seam field but
empty here (the verbatim loop tracks only a count) — T3 populating it is a follow-on.
